refactor(integrations): extract atlassian/ sub-package — move 6 Atlassian files
Move 3 source + 3 test files from internal/tools/integrations/ to
internal/tools/integrations/atlassian/. Update package declaration to
'package atlassian'. Rename atlassian_common.go → provider.go per spec.

Parent integrations/ package and registration.go will be fixed in
subsequent commits (Task A3).
